+++ examples/basic/eth-stream-transfers.ts
@@ -7,12 +7,27 @@ async function main() {
     network: "mainnet.eth.dfuse.io"
   })
 
+  const streamTransfer = `subscription($cursor: String) {
+    searchTransactions(query: "method:'transfer(address,uint256)'", cursor: $cursor) {
+      undo cursor
+      node { hash from to value(encoding: ETHER) }
+    }
+  }`
+
   const stream = await client.graphql(streamTransfer, (message) => {
+    if (message.type === "error") {
+      console.log("An error occurred", message.errors, message.terminal)
+    }
+
     if (message.type === "data") {
       const { cursor, node } = message.data.searchTransactions
       console.log(`Transfer [${node.from} -> ${node.to}, ${node.value}]`)
 
       stream.mark({ cursor })
+    }
+
+    if (message.type === "complete") {
+      console.log("Stream completed")
     }
   })
 
@@ -21,17 +36,5 @@ async function main() {
 
   client.release()
 }
-
-const streamTransfer = `
-  subscription {
-    searchTransactions(query: "method:'transfer(address,uint256)'") {
-      undo
-      cursor
-      node {
-        hash from to value(encoding: ETHER)
-      }
-    }
-  }
-`
 
 runMain(main)